+++ slack-plugin/src/main/kotlin/slack/gradle/SlackBasePlugin.kt
@@ -24,7 +24,6 @@ import org.gradle.api.Project
 import org.gradle.api.artifacts.Configuration
 import org.gradle.api.artifacts.MinimalExternalModuleDependency
 import org.gradle.api.provider.Provider
-import org.jetbrains.kotlin.gradle.plugin.KotlinBasePlugin
 import slack.gradle.tasks.CoreBootstrapTask
 import slack.stats.ModuleStatsTasks
 import slack.unittest.UnitTests
@@ -43,7 +42,11 @@ internal class SlackBasePlugin : Plugin<Project> {
       val versionCatalog =
         target.getVersionsCatalogOrNull() ?: error("SGP requires use of version catalogs!")
       StandardProjectConfigurations(slackProperties, versionCatalog).applyTo(target)
-      UnitTests.configureSubproject(target, slackProperties)
+      UnitTests.configureSubproject(
+        target,
+        slackProperties,
+        target.slackTools().globalConfig.affectedProjects
+      )
       CoreBootstrapTask.configureSubprojectBootstrapTasks(target)
 
       // Configure Gradle's test-retry plugin for insights on build scans on CI only
@@ -60,12 +63,6 @@ internal class SlackBasePlugin : Plugin<Project> {
       if (slackProperties.autoApplyCacheFix) {
         target.pluginManager.withPlugin("com.android.base") {
           target.pluginManager.apply("org.gradle.android.cache-fix")
-        }
-      }
-
-      if (slackProperties.autoApplyDetekt) {
-        target.plugins.withType(KotlinBasePlugin::class.java) {
-          target.pluginManager.apply("io.gitlab.arturbosch.detekt")
         }
       }
 
@@ -190,7 +187,7 @@ internal class SlackBasePlugin : Plugin<Project> {
     checkerDepOptional: Optional<Provider<MinimalExternalModuleDependency>>
   ) {
     val configurationName = configuration.name
-    val lowercaseName = configurationName.toLowerCase(Locale.US)
+    val lowercaseName = configurationName.lowercase(Locale.US)
     // Hamcrest switched to a single jar starting in 2.1, so exclude the old ones but replace the
     // core one with the
     // new one (as cover for transitive users like junit).