+++ slack-plugin/src/main/kotlin/slack/unittest/UnitTests.kt
@@ -28,6 +28,7 @@ import org.gradle.kotlin.dsl.retry
 import org.gradle.language.base.plugins.LifecycleBasePlugin
 import slack.gradle.SlackProperties
 import slack.gradle.ciUnitTestAndroidVariant
+import slack.gradle.configureEach
 import slack.gradle.isActionsCi
 import slack.gradle.isCi
 import slack.gradle.util.synchronousEnvProperty
@@ -67,7 +68,11 @@ internal object UnitTests {
       description = "Global lifecycle task to run all ciUnitTest tasks."
     }
 
-  fun configureSubproject(project: Project, slackProperties: SlackProperties) {
+  fun configureSubproject(
+    project: Project,
+    slackProperties: SlackProperties,
+    affectedProjects: Set<String>?
+  ) {
     // Projects can opt out of creating the task with this property.
     val enabled = slackProperties.ciUnitTestEnabled
     if (!enabled) {
@@ -79,7 +84,19 @@ internal object UnitTests {
       project.pluginManager.apply("org.jetbrains.kotlinx.kover")
     }
 
-    val globalTask = project.rootProject.tasks.named(GLOBAL_CI_UNIT_TEST_TASK_NAME)
+    val globalTask =
+      if (affectedProjects == null || project.path in affectedProjects) {
+        project.rootProject.tasks.named(GLOBAL_CI_UNIT_TEST_TASK_NAME)
+      } else {
+        val log =
+          "$LOG Skipping ${project.path}:$CI_UNIT_TEST_TASK_NAME because it is not affected."
+        if (slackProperties.debug) {
+          project.logger.lifecycle(log)
+        } else {
+          project.logger.debug(log)
+        }
+        null
+      }
 
     // We only want to create tasks once, but a project might apply multiple plugins.
     val applied = AtomicBoolean(false)
@@ -97,7 +114,7 @@ internal object UnitTests {
             group = LifecycleBasePlugin.VERIFICATION_GROUP
             dependsOn("test")
           }
-        globalTask.configure { dependsOn(ciUnitTest) }
+        globalTask?.configure { dependsOn(ciUnitTest) }
         project.tasks.register(COMPILE_CI_UNIT_TEST_NAME) {
           group = LifecycleBasePlugin.VERIFICATION_GROUP
           dependsOn("testClasses")
@@ -110,7 +127,7 @@ internal object UnitTests {
     configureTestTasks(project, slackProperties)
   }
 
-  private fun createAndroidCiUnitTestTask(project: Project, globalTask: TaskProvider<*>) {
+  private fun createAndroidCiUnitTestTask(project: Project, globalTask: TaskProvider<*>?) {
     val variant = project.ciUnitTestAndroidVariant()
     val variantUnitTestTaskName = "test${variant}UnitTest"
     val variantCompileUnitTestTaskName = "compile${variant}UnitTestSources"
@@ -122,7 +139,7 @@ internal object UnitTests {
         // task configuration time.
         dependsOn(variantUnitTestTaskName)
       }
-    globalTask.configure { dependsOn(ciUnitTest) }
+    globalTask?.configure { dependsOn(ciUnitTest) }
     project.tasks.register(COMPILE_CI_UNIT_TEST_NAME) {
       group = LifecycleBasePlugin.VERIFICATION_GROUP
       // Even if the task isn't created yet, we can do this by name alone and it will resolve at
@@ -135,7 +152,7 @@ internal object UnitTests {
     val isCi = project.isCi
 
     // Unit test task configuration
-    project.tasks.withType(Test::class.java).configureEach {
+    project.tasks.configureEach<Test> {
       // Run unit tests in parallel if multiple CPUs are available. Use at most half the available
       // CPUs.
       maxParallelForks =