+++ CHANGELOG.md
@@ -1,6 +1,142 @@
 Changelog
 =========
 
+0.7.9
+-----
+
+_2023-04-01_
+
+Happy April Fool's Day!
+
+- [Skippy] Recursively resolve project dependencies to avoid missing transitive edges in the graph. Previously we only computed shallow dependencies.
+
+0.7.8
+-----
+
+_2023-03-28_
+
+- **Fix**: Add missing `detekt` task dependencies for `globalDetekt`.
+- **Fix**: Only apply detekt config once (even if multiple Kotlin plugins are applied).
+
+0.7.7
+-----
+
+_2023-03-27_
+
+- Add new `slack.detekt.full` property to gate whether or to run full detekt (i.e. with type resolution). If disabled, `detektRelease`/`detektMain` and associated tasks will be disabled and not used in `detektGlobal`.
+
+0.7.6
+-----
+
+_2023-03-25_
+
+- **Fix**: Apply matching configurations to `DetektCreateBaselineTask` tasks too due to https://github.com/detekt/detekt/issues/5940.
+
+0.7.5
+-----
+
+_2023-03-24_
+
+- [Skippy] Add more default configurations.
+- [Skippy] Add `slack.avoidance.build-upon-default-affected-project-configurations` flag to make provided configurations build upon defaults.
+- Add new `globalDetekt` task that runs `detekt` on all subprojects. This is Skippy-compatible and responds to `slack.avoidance.affectedProjectsFile`.
+
+0.7.4
+-----
+
+_2023-03-22_
+
+- Don't expose `androidExtension` publicly in `SlackExtension` to avoid Gradle mismatching number of type arguments in AGP 8.1.0-alpha10+.
+
+0.7.3
+-----
+
+_2023-03-22_
+
+- Set `Detekt.jdkHome` to null to avoid https://github.com/detekt/detekt/issues/5925.
+- Rename `String.safeCapitalize()` to `String.capitalizeUS()` to make it more explicit.
+
+0.7.2
+-----
+
+_2023-03-21_
+
+- Disable Live Literals in Compose by default due to multiple issues. They can be enabled via `-Pslack.compose.android.enableLiveLiterals=true`.
+  - Poor runtime performance: https://issuetracker.google.com/issues/274207650.
+  - Non-deterministic class files breaking build cache: https://issuetracker.google.com/issues/274231394.
+- [Skippy] Add `.github/actions/**` to default never skip filters.
+
+0.7.1
+-----
+
+_2023-03-20_
+
+- [Skippy] Improve pattern configuration.
+  1. Make the default patterns public. This allows consumers to more
+     easily reuse them when customizing their own.
+  2. Use sets for the type to better enforce uniqueness requirements.
+  3. Add github actions to never-skip defaults.
+  4. Add excludePatterns to allow finer-grained control. This runs _after_
+     include filtering so that users can manually exclude certain files that
+     may otherwise be captured in an inclusion filter and is difficult to
+     describe in a simple glob pattern. GitHub action does similar controls
+     for CI matrices.
+- [Skippy] Allow relative (from the project root) to `affected_projects.txt` and allow non-existent files as a value. This makes it easy to gracefully fall back in CI.
+- [Skippy] Fix logging path matchers missing toString() impls.
+- [SKippy] Log verbosely in debug mode when skipping task deps.
+- Update oshi to `6.4.1`.
+
+0.7.0
+-----
+
+_2023-03-17_
+
+### Project Skippy
+
+This release introduces an experimental new `computeAffectedProjects` task for computing affected projects based on an input of changed files. The goal of this is to statically detect which unit test, lint, and androidTest checks can be safely skipped in CI on pull requests.
+
+Example usage
+```bash
+./gradlew computeAffectedProjects --changed-files changed_files.txt
+```
+
+Where `changed_files.txt` is resolved against the root repo directory and contains a newline-delimited list of changed files (usually inferred from a PR).
+
+A simple example of how to produce such a file with the `gh` CLI:
+
+```bash
+gh pr view ${{ github.event.number }} --json files -q '.files[].path' > changed_files.txt
+```
+
+One would run this task _first_ as a preflight task, then run subsequent builds with the `slack.avoidance.affectedProjectsFile` Gradle property pointing to its output file location (printed at the end of the task).
+
+```bash
+./gradlew ... -Pslack.avoidance.affectedProjectsFile=/Users/zacsweers/dev/slack/slack-android-ng/build/skippy/affected_projects.txt
+```
+
+The `globalCiLint`, `globalCiUnitTest`, and `aggregateAndroidTestApks` tasks all support reading this property and will avoid adding dependencies on tasks in projects that are not present in this set.
+
+The `ComputeAffectedProjectsTask` task has some sensible defaults, but can be configured further in the root projects like so.
+
+```kotlin
+tasks.named<ComputeAffectedProjectsTask>("computeAffectedProjects") {
+  // Glob patterns of files to include in computing
+  includePatterns.addAll(
+    "**/*.kt",
+    "**/*.java",
+  )
+  // Glob patterns of files that, if changed, should result in not skipping anything in the build
+  neverSkipPatterns.addAll(
+    "**/*.versions.toml",
+    "gradle/wrapper/**",
+  )
+}
+```
+
+Debug logging can be enabled via the `slack.debug=true` Gradle property. This will output timings, logs, and diagnostics for the task.
+
+The configurations used to determine the build graph can be customized via comma-separated list to the `slack.avoidance.affected-project-configurations` property.
+
 0.6.1
 -----